Moved logic to VerifiableCredentialExtractor.
