+++ extensions/identity-hub-verifier/src/main/java/org/eclipse/dataspaceconnector/identityhub/verifier/CredentialsVerifierExtension.java
@@ -18,6 +18,7 @@ import okhttp3.OkHttpClient;
 import org.eclipse.dataspaceconnector.iam.did.spi.credentials.CredentialsVerifier;
 import org.eclipse.dataspaceconnector.iam.did.spi.resolution.DidPublicKeyResolver;
 import org.eclipse.dataspaceconnector.identityhub.client.IdentityHubClientImpl;
+import org.eclipse.dataspaceconnector.identityhub.credentials.VerifiableCredentialExtractor;
 import org.eclipse.dataspaceconnector.spi.EdcException;
 import org.eclipse.dataspaceconnector.spi.EdcSetting;
 import org.eclipse.dataspaceconnector.spi.monitor.Monitor;
@@ -65,6 +66,7 @@ public class CredentialsVerifierExtension implements ServiceExtension {
         }
 
         var client = new IdentityHubClientImpl(httpClient, typeManager.getMapper(), monitor);
-        return new IdentityHubCredentialsVerifier(client, monitor, jwtCredentialsVerifier);
+        var verifiableCredentialExtractor = new VerifiableCredentialExtractor(typeManager.getMapper());
+        return new IdentityHubCredentialsVerifier(client, monitor, jwtCredentialsVerifier, verifiableCredentialExtractor);
     }
 }
