+++ extensions/identity-hub-verifier/src/main/java/org/eclipse/dataspaceconnector/identityhub/verifier/IdentityHubCredentialsVerifier.java
@@ -17,6 +17,7 @@ package org.eclipse.dataspaceconnector.identityhub.verifier;
 import org.eclipse.dataspaceconnector.iam.did.spi.credentials.CredentialsVerifier;
 import org.eclipse.dataspaceconnector.iam.did.spi.document.DidDocument;
 import org.eclipse.dataspaceconnector.identityhub.client.IdentityHubClient;
+import org.eclipse.dataspaceconnector.identityhub.credentials.VerifiableCredentialExtractor;
 import org.eclipse.dataspaceconnector.spi.monitor.Monitor;
 import org.eclipse.dataspaceconnector.spi.result.AbstractResult;
 import org.eclipse.dataspaceconnector.spi.result.Result;
@@ -35,16 +36,18 @@ public class IdentityHubCredentialsVerifier implements CredentialsVerifier {
     private final IdentityHubClient identityHubClient;
     private final Monitor monitor;
     private final JwtCredentialsVerifier jwtCredentialsVerifier;
+    private final VerifiableCredentialExtractor verifiableCredentialExtractor;
 
     /**
      * Create a new credential verifier that uses an Identity Hub
      *
      * @param identityHubClient IdentityHubClient.
      */
-    public IdentityHubCredentialsVerifier(IdentityHubClient identityHubClient, Monitor monitor, JwtCredentialsVerifier jwtCredentialsVerifier) {
+    public IdentityHubCredentialsVerifier(IdentityHubClient identityHubClient, Monitor monitor, JwtCredentialsVerifier jwtCredentialsVerifier, VerifiableCredentialExtractor verifiableCredentialExtractor) {
         this.identityHubClient = identityHubClient;
         this.monitor = monitor;
         this.jwtCredentialsVerifier = jwtCredentialsVerifier;
+        this.verifiableCredentialExtractor = verifiableCredentialExtractor;
     }
 
     /**
@@ -68,7 +71,7 @@ public class IdentityHubCredentialsVerifier implements CredentialsVerifier {
                 .filter(jwt -> jwtCredentialsVerifier.verifyClaims(jwt, didDocument.getId()))
                 .filter(jwtCredentialsVerifier::isSignedByIssuer);
 
-        var credentials = verifiedClaims.map(VerifiableCredentialUtil::extractCredential).collect(Collectors.toList());
+        var credentials = verifiedClaims.map(verifiableCredentialExtractor::extractCredential).collect(Collectors.toList());
         credentials.stream().filter(AbstractResult::failed).forEach(result -> monitor.warning(String.join(",", result.getFailureMessages())));
 
         var claims = credentials.stream().filter(AbstractResult::succeeded)
